refactor: P1-T3 - 抽出ChatDetailView保持现状
